Drain barrier worklist into mark queue if it is too large
Summary:
We currently allow the write barrier worklist to grow indefinitely if no
OG marking takes place. This means that if there are no allocations, the
write barrier worklist could easily grow large enough to crash the
process.

To address this, have the write barrier slow path check if the list has
grown too large, and force the list to be drained on the mutator.

Reviewed By: lavenzg

Differential Revision: D43551913

fbshipit-source-id: d0744d40b4fcbc4c75f5c43ebb3dbcf9ceeca190