VM: Simplify toArrayIndex operations
Summary:
toArrayIndex shouldn't need to create a `StringView` and take a
`Runtime`. Therefore it doesn't need a `Handle`.
Just use `StringPrimitive *` for that overload, but leave the
`StringView` overload as well because that works for lazy identifiers
without having to materialize them.

This doesn't yield any speedup from what I can tell, it's just a
simplification.

Reviewed By: neildhar

Differential Revision: D77396388

fbshipit-source-id: 11537bb8233718a95c049b93805d3084dc591d5c