Use new debug info in Debugger::evalInFrame
Summary:
Reimplement `Debugger::evalInFrame` to use the new debug info. This
information informs the debugger on where to find the environment
holding the variables of the nearest-enclosing scope, and the
LexicalScope which describes that nearest-enclosing scope.

This also means that in order to find this information, we need to
inspect which bytecode instruction exactly we are stopped on. Previously
the information was not this fine-grained, so we only needed to inspect
function-level information.

There are also lots of minor code changes needed to pass the
LexicalScope and VariableScope through back to the compiler when
restarting compilation.

Reviewed By: avp

Differential Revision: D74186405

fbshipit-source-id: afba637c87fe7e3c025525660ec1a97607199138