Track implicit environment operand for all instructions
Summary:
Associate an implicit environment operand with all instructions. This is
done by adding a new `uint32_t environmentId_` field to `Instruction`.
This
integer ID can be interpreted to mean 3 different kinds of values:

- 0 means no env info is associated with the instruction
- 1 means the current parent environment
- =>2 is an index into a list of environments

This list of environments is owned by an IR `Function`. Each time a
scope is created in IRGen (e.g. `CreateScopeInst`), a new element is
pushed into this list.

The current environment index is owned by the IR `Function` as well.
This was done to match how `statementCount_` works. By putting it into the
`Function`, it becomes easier to manage the lifetime of the env index. In the
future we may want to move this state into `IRBuilder`, but that would require
some more careful thought around initializing this state correctly and
restoring it when exiting functions.

Future diffs will extend this list to also be able to contain `Variable`s
for spilled environments.

This functionality is only triggered with full debug info.

Reviewed By: avp

Differential Revision: D75538940

fbshipit-source-id: e251122320b1d3a0301f5e2a150df814d611a372