Add new debug info for location of environments and lexical scope
Summary:
Extend the bytecode file format to associate with each BC instruction a
pair of `<env location, lexical scope>`.

The lexical scope is simply a pointer to a `sema::LexicalScope`. This
will be changed in the future to an ID, but this is the simplest
incremental step towards working debug info. The environment location
can express one of 3 kinds of places: the parent environment, a
register, or a slot in the parent environment.

Since we only support debugging from source, we only care to add the
side table directly in memory to `DebugInfo`. The file format itself is
then updated to just hold an additional field which is an index into
this side table. We modify the format because it's the most convenient
way of expressing information per-BC instruction.

In the future, we will modify the lexical scope to be an integer ID
instead of a pointer. This will allow for a high rate of global
de-duplication of these entries. As in, these entries can be de-duped
across functions. In preparation for this, we also modify `ISel` to take
in the `DebugInfoGenerator` directly.

Reviewed By: avp

Differential Revision: D74186402

fbshipit-source-id: fcfc1e677b30a57981b1679f7184b8abee52ba94