Preserve implicit environment operand
Summary:
From each instruction, we can now find the nearest enclosing
`CreateScopeInst`. During register allocation, we need to act as if this
instruction is an operand, to extend the lifetime of it. This way, we
ensure that the environment we need to access at runtime is always
available.

Defensively, we add a new pass to check for dominance. If this implicit
operand does not dominate its instruction, then we set the environment
ID of the instruction to 0, representing no information. This can happen
because of lowering generators, where the structure of `yield`s can
result in this structure.

Reviewed By: neildhar

Differential Revision: D73611600

fbshipit-source-id: 56bc8df3852940fed784dcd89244acc055b3ca2a