Support BigInt property in object literals
Summary:
BigInts are allowed as property keys because the spec thinks of them as
a type of NumericLiteral. Like doubles, they're just converted to
strings when used as keys.

Now that the `bigint` AST field is the string representation of the
BigInt, it's easy to pass that information along through IRGen.

Fix https://github.com/facebook/hermes/issues/1725

Reviewed By: fbmal7

Differential Revision: D77455610

fbshipit-source-id: 9ebacc8ca3a75f967b44d0bcb8ce52c192377e85