Parser: Don't include '_' and 'n' in bigint literals
Summary:
ESTree spec says that

> bigint property is the string representation of the BigInt value. It
> must contain only decimal digits and not include numeric separators
> (_) or the suffix n.

So modify the parser to filter those out and no longer store the raw
string.

This allows simplifications in the BigInt NumericLiteral parsing code
because it no longer needs to handle the trailing `n`.

The JS side of this (hermes-parser) was already filtering out `_` and
`n` before converting to the JS AST so there's no changes required
there.

Reviewed By: tmikov

Differential Revision: D77455612

fbshipit-source-id: 527d791e06d9cbfe566a610a89c350d815871677