Check for generators in class accessors to allow for ASI
Summary:
When parsing the class
```
class A {
  get
  *a() {}
}
```
we need to account for the fact that `get *` will result in a regular
`get` property followed by ASI. Handle this by also checking for `*`
and treating the `get` as a field.

Do the same for `set`.

Fixes https://github.com/facebook/hermes/issues/1726

Reviewed By: tmikov

Differential Revision: D77544960

fbshipit-source-id: 03f1efae7d3e50d66ad40fc81f206eaec853d17d