Close generators that throw and have no try/catch
Summary:
Previously, we incorrectly attempted to perform a small optimization
when lowering generators. We would only install our synthetic try/catch
structure if the user had no `try`s in the code. This almost works, in
the sense that this does result in the behavior that a thrown exception
simply bubbles out of the generator and is not caught at all.
However, this does have the unintended side effect of not modifying the
generators's internal state correctly. Specifically, the generator
should now be marked as in the 'completed' state. In order to do this,
we must always guard the entire user code in a try/catch. This way we
can ensure we are always catching the moment when an exception bubbles
out of the generator, and set the state to complete.

Reviewed By: avp

Differential Revision: D77600896

fbshipit-source-id: 062a42636880353f89c4dc6640a679ec07bd1a23