Add caching for getVariableInfoAtDepth
Summary:
In order to perform a search to find `variableIdx` on a scope level, we
need to iterate through the entire `decls` array. This is because the
reported amount of variables is filtered to avoid decls like private
names. So, `variableIdx` practically means "index into the list of decls
that pass the filter." Since this list isn't actually stored anywhere,
we are essentially recreating each query by counting the number of
filtered decls and stopping when we've reached the desired amount.
Instead of having to start this search from scratch each time, we add a
simple layer of caching. We cache the last lexical scope and actual
index into the `decls` array that we found the last queried
`variableIdx`. We can then start from where we left off when we query
the next `variableIdx+1`, which is a calling pattern that happens when
listing out all variables in the debugger.

Reviewed By: avp

Differential Revision: D77670935

fbshipit-source-id: 33081ccdc540a8ceece86022201e3f66d061ec83