Fix printing variable
Summary:
Make printing variables work again with the new debug info. We need to
change what the concept of a 'scope' means throughout `Debugger.cpp`/
`DebuggerAPI.h`. Instead of iterating through `VariableScope`s, we
instead iterate through the chain of `LexicalScope`s. We then need to do
some additional work to get back to the `VariableScope`s and the index
into each `VariableScope`, since ultimately that is the information
that's needed to read a variable out of an environment. We do that by
looking into the customData for `LexicalScope` and `Decl`, which store
where to find the `VariableScope` and the index in that VariableScope.

TODO: next diff we will add some caching to avoid having to iterate
through potentially all of `decls` each time we look up a variable at
`variableIdx`.

Reviewed By: avp

Differential Revision: D77405251

fbshipit-source-id: 3d2a624779e07d569a24ae6b7c1175e1cda6526f