Make DummyRuntime Support Locals and LocalsRAII
Summary:
We use `DummyRuntime` in a lot of unit tests, which does not support
`Locals` and `LocalsRAII`. So currently we have to use GCScope in such
tests. This diff adds the `vmLocals` field and marking of its values
into `DummyRuntime`. As an example, migrate `GCScope` and `Handle` to
`Locals` and `PinnedValue` in `LOABasicOperations()`. I will migrate
the rest `GCScope` in unit tests in a separate diff.

Reviewed By: fbmal7

Differential Revision: D77876411

fbshipit-source-id: 51f6e3dc2e0040e446400064c26d6e94f28a475e