VM: Resize JSArray storage to requested length in NewArray
Summary:
NewArray with a nonzero length will likely be followed with instructions
that will populate the elements from `0` to the length.
Make this faster by immediately resizing the underlying `ArrayStorage`
and filling it with `empty`s (and updating the `elemCount` of the
JSArray).

M4 getNodeById goes from 38.4 B to 37.6 B instructions.
3% perf improvement on draw, nothing elsewhere.

Reviewed By: lavenzg

Differential Revision: D77560788

fbshipit-source-id: 9182ab734f8ce2a26f875817abfd34cb887b7f27