Add a DefineOwnInDenseArray instruction
Summary:
Prior to this diff, we used `DefineOwnByIndex` to populate array
literals after they were created (if the elements were not themselves
literals). But doing that incurred extra cost due to checks for whether
the write was allowed, what the size of the storage was, etc.

Now that `NewArray` resizes the underlying storage to accommodate all
the elements, we can have a much faster instruction. It just uses
`unsafeSetExistingElementAt` and doesn't need to perform any checks
outside of encoding the `SmallHermesValue`.

M4 results for getNodeById:
insts: 37.6 B -> 36.0 B

Pi 5 shows wins on draw and pixel-grid-small:
```
+------------------+------------+-----------------+------------+--------------+
| Benchmark        | Avg Change |    95% interval |      Stdev |   Valid Runs |
+------------------+------------+-----------------+------------+--------------+
| draw             |      +2.0% |   +0.9% - +3.1% |   +/- 1.1% |     (5 runs) |
| pixel-grid       |      +1.8% |   -0.2% - +3.8% |   +/- 2.0% |     (5 runs) |
+------------------+------------+-----------------+------------+--------------+
```

Results are just for this diff.

NOTE: The changes in StackTracesTreeTest are because the new instruction
can't throw so extra source locations aren't emitted for memory
instrumentation to use.

Reviewed By: fbmal7

Differential Revision: D77470892

fbshipit-source-id: c1d062908da507a1962ba6e0f4a6ff95e7e280f2