Unique the DebugScopingInfo side table
Summary:
Now that `DebugScopingInfo` stores an integer ID for the lexical scope, it
is essentially just a `std::pair` of `<int64_t, uint32_t>`. These values
should have a high rate of duplication both inside of a single function
but also even across functions. We take advantage of this expected
similarity now and add de-duplication logic to `DebugInfoGenerator`.

Reviewed By: avp

Differential Revision: D77945943

fbshipit-source-id: aeadf839771931fbced697820db94235ffd752f6