Visit super class node before declaring private names
Summary:
The super class node (the expression in the `extends`) must be visited
before we collect and declare the private names. This is because that
expression should not be allowed to use private names declared in the
body of the class.

Now that we are preemptively visiting one of the children of the class
node, we can't use `visitESTreeNode` since that would then revisit the
same super class node. So make a small helper to visit the non-super
class node children and call it where we previously used
`visitESTreeNode`.

Reviewed By: avp

Differential Revision: D78028211

fbshipit-source-id: 404c35503b47208048af44c0cca1c460ce547c09