Fix deltamode handling of new identifiers
Summary:
1. The `populateStringsTableFromStorage` function wasn't handling having
multiple strings with the same contents correctly. It needs to store the
new index, because the new index can be an identifier, and the value in
`strings_` is supposed to be the most recent string ID for that string.
2. The `count()` and `empty()` functions were checking `strings_`
instead of the actual table. `strings_` can't have duplicates and
therefore it undercounts the actual number of string table entries.

Reviewed By: dannysu, lavenzg

Differential Revision: D78136810

fbshipit-source-id: aac1ba782049d89e6a5f5bcf4e4f71e63c4d6c47