Fix internal Hermes breakpoint validation
Summary:
When Debugger domain is enabled, we only have to validate that there are no internal Hermes breakpoints attached to the CDP Breakpoints.

Before this change, because we are running this assertion for every possible pair of Source Script and CDPBreakpoint, it would have failed in cases where there the internal breakpoint was applied for one of the first loaded scripts. More concretely, it would've failed on the next iteration step in outer for loop.

Reviewed By: dannysu

Differential Revision: D78165903

fbshipit-source-id: 50d801bd1597e104b5c8c8b726ff05f2b227aee8