Delete assert in ESTreeIRGen::genScopedForLoop
Summary:
`genScopedForLoop` is invoked when there is a *chance* that we need the
specialized code path for setting up variables in new scopes. However,
it is easy to get a false positive from this heuristic, and call into
`genScopedForLoop` when we actually didn't need to, and there are no
variables to be created. E.g.:

```
for (const [] = []; class C {};) {
  break;
}
```

So we should remove the assert that mandates there is at least one
`Decl` present in the for loop's `sema::LexicalScope`.

Reviewed By: avp

Differential Revision: D78838906

fbshipit-source-id: 836f137d68c28395028d43a34ebf87e45daf4e22