Add event loop to shermes
Summary:
Handle Promise and setTimeout properly in the `shermes` tool with the
native backend by expanding the functionality of the shermes
ConsoleBindings.

Allow `init_console_bindings` to store some extra data that can be kept
around after the unit is initialized and run, and have a separate
function for running the event loop.

The logic is largely implemented in JS, making it pretty simple to
understand.

Reviewed By: lavenzg

Differential Revision: D78506822

fbshipit-source-id: 2e6e66104995a64831e9153f4e36ad7f01063cb4