Use fence to prevent reordering in concurrentReadImpl
Summary:
`atomic_signal_fence` is a function that prevents the compiler from
reordering instructions across the fence.

Instead of type casting to a volatile pointer and using a `union` to
convert back and forth between a `T` and an integer, copy `T` directly
(which is now possible because `HermesValue` now has a copy ctor)
and use a fence to prevent the compiler moving the read later.

This results in a much simpler version of the function.

Reviewed By: tmikov

Differential Revision: D79121959

fbshipit-source-id: a341cae03757f1bbd76250ccfa5033e390010995