Skip lazy functions when running async generators transform
Summary:
Much like with `SemanticResolver`, the AST transform should not descend
into the bodies of async generator functions if they are lazy.
This leads to messing with the structure which we rely on to know
what to parse when we actually call the lazy function.

To fix this, skip any functions with a lazy body when running the AST
transform.

However this has a side effect - the rest of the pipeline needs to
handle the lazy async generator function. Deal with this by
updating the check in `SemanticResolver` to check for the async
generators flag explicitly.
Remove the check in IRGen because `SemanticResolver` is
already checking this.

This incidentally worked for ordinary function declarations/expressions
because the generator created by the transform was still lazy. However,
it failed for async generator methods, because we set their locations
based on their parent AST nodes (for reparsing) and the parent AST node
is modified by the async generator transform.

Reviewed By: lavenzg

Differential Revision: D79201367

fbshipit-source-id: 992c431cc09d66dd4c8e139d6cbd3740d765a085