Set non-zero sanitize rate in DummyRuntime
Summary:
It's useful to have non-zero sanitize rate in unit tests. This diff
sets it to 0.01, which is the same as the hermes standalone tool.
This helped uncover a few bugs in unit tests (e.g., SymbolID not
preserved in a GC aware locations, allocating an object as argument
without considering the function call evaluation order).

Reviewed By: avp

Differential Revision: D78301418

fbshipit-source-id: 765f1517d1ce919f93fd79dc1c3519e82a48cab3