Enable ARM32 test on Sandcastle with cross-compilation
Summary:
This diff adds a shell script that cross-compiles SH to ARM32 and run
tests. As a first step, a few tests are skipped either because we need
to find a way to add the `qemu-arm` prefix or they failed in qemu user
mode. The test262 test will be added later.

Reviewed By: avp

Differential Revision: D79204231

fbshipit-source-id: 263b78caebdaa8db6604de90f6a017a3f9088bd9