Parser: Fix grammar context after type args in JSX
Summary:
In JSX, more identifiers are allowed than in Flow or JS, because
identifiers can contain `-`.
We handle this with `JSLexer::GrammarContext::AllowJSXIdentifier`,
which has to be passed in to lex JSX identifiers correctly.

JSX openings can contain Flow type arguments:
```
<A<T> foo-bar=1 />
```
which are parsed by `parseTypeArgsFlow()`.

However, because `parseTypeArgsFlow` also advances past the final `>` in
the type arguments (after the `T` above), it needs to specify the
correct `GrammarContext` in order to ensure that the `foo-bar` attribute
has the entirety of `foo-bar` lexed as a single identifier.

Handle this by allowing the caller to specify the trailing grammar
context of the type arguments.

Reviewed By: tmikov

Differential Revision: D80293212

fbshipit-source-id: af791540598f1a104e9f74bbacdb3c18bbdf75e6