Fix a few unit tests
Summary:
These were uncovered when running tests with 64KB segment size. There
are mainly two categories of failures:
1. A few JSON parsing unit tests failed because
`StringPrimitive::createWithKnownEncoding()` may return ASCII string if
it's not external, and UTF16 if it's external. With small segment size,
the string in the test becomes external.
2. A few GC related tests have wrong computation of maximum memory to
allocate. FixedSizeHeapSegment::maxSize() returns the maximum
allocation size, not the size of memory for that segment.

Reviewed By: tsaichien

Differential Revision: D79930745

fbshipit-source-id: 938e7c7267b0f275c1d5b6ed98b0e1f86d0879a2