Move StackFrameLayout into a C header
Summary:
In order to more easily read parameters, we need to be able to
understand the stack frame layout from the native backend.
Move the field enum contents and documentation to the C header,
and use it to populate the C++ version, which is unchanged.

In doing this, move the stack frame layout information to a separate
`include` directory called `VMLayouts`.

Reviewed By: tmikov

Differential Revision: D78853215

fbshipit-source-id: a3dba10b1f0cdc3d7064d4970eaed993179723dd