Rename `StripComponentSyntax` to `TransformComponentSyntax`
Summary:
`StripComponentSyntax` was probably named because of how `StripFlowTypes` was named (the transform that came before it),
however unlike `StripFlowTypes` which just removes information (stripping it out), the component syntax transforms its syntax into something else.
Also, this will be in line with the other syntax transforms like `TransformMatchSyntax`, and the future `TransformEnumSyntax`.

Changelog: Update any imports of `StripComponentSyntax` to `TransformComponentSyntax`

Reviewed By: SamChou19815

Differential Revision: D80851424

fbshipit-source-id: 907dd2ce0932537fb9a3f4fc26b3ff05ea331b43