Reorganize printHeapStats
Summary:
Currently Runtime::printHeapStats() is a bit messed. Various stats are
printed out, but this function is called if "--gc-print-stats" is
specified. Move call of getIOTrackingInfoJSON to ConsoleHost,
conditioned on "--track-io". Check "--gc-print-stats" in
printArrayCensus().

With this change, we should be able to call printHeapStats() in the
shermes backend, without checking the flag there.

Reviewed By: avp

Differential Revision: D81510406

fbshipit-source-id: 09bd0a3ad5a3480f53e783d4725842481cb55a21