Fix `setPropertyValue` and `deleteProperty` consistency
Summary:
Both `setPropertyValue` and `deletePropeety` pretends that it is on
strict mode and should throw on error. JSI default implementation does
the correct thing. For the Hermes implementation, the initial change
missed adding the `throwOnError` flag for `setPropertyValue`. For
`deleteProperty`, `throwOnError` flag is already used, but we shouldn't
need to check on the result of the `deleteComputed`.

Reviewed By: lavenzg

Differential Revision: D81342175

fbshipit-source-id: d0f11fab7f519c7c2a765e4c3750e0556e715641