Fix `setPropertyValue` and `deleteProperty` consistency
Summary:
Both `setPropertyValue` and `deletePropeety` pretends that it is on
strict mode and should throw on error. JSI default implementation does
the correct thing. For the Hermes implementation, the initial change
missed adding the `throwOnError` flag for `setPropertyValue`. For
`deleteProperty`, `throwOnError` flag is already used, but we shouldn't
need to check on the result of the `deleteComputed`.

Grafted 5519eeb956367c52005906ebd13700dbc3090da7
- Grafted xplat/static_h to xplat/hermes

Reviewed By: lavenzg

Differential Revision: D81342263

fbshipit-source-id: 05eba0f7f6fefa7f191d62a6c793dc83750abf06