Fix prefab config for publishing Hermes
Summary: The current prefab config sets the `libraryName` to `hermesvm`, which causes issues when consuming it. The consumer ends up looking for `hermesvm.so` instead of `libhermesvm.so`. This is a leftover from the previous naming scheme where the module was named `libhermes` and the override was necessary to avoid double `lib` prefix.

Reviewed By: huntie

Differential Revision: D81768434

fbshipit-source-id: 90345fff87179faab04b9e1fdc01fc6ee2cba26f