IRGen: Use Phi for optional chaining 'this' value
Summary:
In the expression
```
(a?.b.c)();
```
The member expression in the parentheses needs to report a `this` value
for use in the call expression.

Bug: The result of loading `b` from `a` was always used as the
`this` value returned from `genOptionalMemberExpression`, but that value
is generated in one branch of the optional chain so it doesn't dominate
the `continueBB` at the end of the optional chain. In practice, this
code will necessarily throw if the optional chain short circuits because
you can't call `undefined`, but the invalid IR must be fixed.

Fix: Use a Phi instruction for the `baseValue` in the way that we use a
Phi for the result value. It's `undefined` for the short circuit block,
because the call will be to `undefined` and will fail.

Fix #1768.

Reviewed By: fbmal7

Differential Revision: D80946292

fbshipit-source-id: 4daa6fcf79d00fc1b70ffe802124510738c11976