Fix generator Phi lowering
Summary:
`moveInnerPhisToOuter` processes Phi instructions in the
`LowerGeneratorFunction` pass.
In order to not move Phis unnecessarily, it uses `shouldMovePhiInst`.

`shouldMovePhiInst` was only looking at whether the block operands to
the Phi were no longer predecessors in order to determine whether to
move the Phi, but that's insufficient. It is possible for the value
operand to the Phi to be in a different block than the block operand
itself, and for there to be a potential `yield` between the value and
the Phi.

The test case is
```
function *foo(f, a, b) {
  "noinline";
  a = [];
  if (f) a = yield 1;
  a = f ? a : b;
  // 'a' is potentially the original [].
  print(a);
}
```
in which the `a` used in the `print` call is a Phi between the two
blocks after the yield, but one of the values is the original
AllocArray.

We need to actually check that the Phi value still dominates the
Phi instead of checking that the block operand still dominates the Phi.

Reviewed By: fbmal7

Differential Revision: D80975720

fbshipit-source-id: 366b8e0222671e655b66d953f447c73ddd9a4eb3