Guard against stack overflow when dumping AST
Summary:
When dumping the AST, we currently do not check depth while traversing
the Nodes. This can results in a stack overflow if the traversal is too
deep (due to a long function chaining, for example).

This diff adds a simple guard against stack overflow.

If `HERMES_CHECK_NATIVE_STACK`, then it will use real stack checking,
checking the current stack address against the upper bound of the stack
and some defined buffer. The current buffer is set at 512 * 102.

Otherwise, it will simply increase the depth as the dumper traverse
through the node, and check the depth against a set limit. The current
limit is set at 128.

Reviewed By: fbmal7

Differential Revision: D81635300

fbshipit-source-id: 65a9446250b01d2d7f4609264296f8a0220ddbc6