Distribute version across `rn-build-hermes` workflow (#1781)
Summary:

Adds distribution of the hermes version across `rn-build-hermes` workflow. This is necessary because on commitly, the Maven Android artifacts are published on the `build_android` workflow, but iOS artifacts and npm package are published later in the `publish` workflow. The `getVersion` method creates a version for Commitly based on the date, which depends on the time at which it was invoked, differing for `build_android` and `publish` workflows.

For instance, latest version for iOS artifacts is
```
0.14.0-commitly-202509121544-0fc5b0fe4-SNAPSHOT
```

and for Android artifacts:
```
0.14.0-commitly-202509121528-0fc5b0fe4-SNAPSHOT
```

which differ in the date part.

This also fixes updating `hermes-compiler/package.json` version for commitlies.

Differential Revision: D82432172