Move flags earlier in Hades for faster checking
Summary:
In order to quickly load from the `xRuntime` register in the JIT,
these flags are better placed higher up in the Hades GC structure.

If we add just a few fields to `Runtime`, it moves the
`ogMarkingBarriers_` flag out of range of the ARMv8 load instruction,
requiring use of a temporary register.
Putting commonly accessed fast path flags higher up prevents the issue.
I moved a couple other flags as well to keep these single-byte fields
together to avoid unnecessarily increasing the size of `HadesGC`.

Also update the JIT to use `emit_load_from_base_offset` with
`INLINE_LOAD=true` for consistency in how we load in the JIT.

Reviewed By: lavenzg

Differential Revision: D82168607

fbshipit-source-id: e4e5eb11011cf589d96989b648b7b7554dc3fb91