Add missing assert to JSObject::getParent()
Summary:
This proxy assert should be in both methods:
getParent and getParentGCPtr.
Use one function in the other to include the assertion.

This exposed a bug in `JSError` that we were already testing for in
`regress-proxy-parent.js`, so the changes to the
`getErrorFromStackTarget` are necessary to fix
the assertion failure on that test.

Reviewed By: tmikov

Differential Revision: D82684111

fbshipit-source-id: ab694692fc2fb19278c79a10ae1aeffe7b41ea47