Copy key buffer and shape table under `base-bytecode`
Summary:
Similar to the value buffer from last diff, copy the key buffer and
shape table from the base bytecode. This ensures that small changes to
the JS source that affect these buffers do not result in changes to the
bytecode instructions that reference the same shapes from the base
bytecode.

The shape table functions as a header for the key buffer, so it's easy
to reconstruct the individual elements in the key buffer.

Reviewed By: avp

Differential Revision: D81496655

fbshipit-source-id: 85425441a5c8cdad94297eff196c56e0454cd0b7