Copy values buffer under `base-bytecode`
Summary:
The motivation for this diff (and the rest of the stack) is to improve
'delta mode', as in when compiling with a given `-base-bytecode`. The
goal is to be more resilient to minor tweaks in the given source.

An optional bytecode provider is given to
`LiteralBufferBuilder::generate`. If this is present, then the new value
buffer created will be initialized as an identical copy of the value
buffer in the base bytecode. This means that even if a new set of values
is used in a literal at the beginning of the new JS source, it will not
displace the indices of all the other value streams that are the same as
in the original bytecode.

There is no convenient header in the bytecode file that describes the
layout of the value buffer. Instead, the individual pieces of the buffer
are reconstructed by inspecting the operands of bytecode instructions
that use the value buffer. So all the instructions in the base bytecode
are iterated, and all the different value buffer usages are collected
and uniqued.

Reviewed By: avp

Differential Revision: D81496654

fbshipit-source-id: d84066c50182e3b1f9bc9d59edb34e3b1c555e38