Add new iterator functions
Summary:
This diff adds new functions for implementing iterators:
- newIterator
- advanceIterator
- iteratorKey
- iteratorValue

The way these functions work is different than the previous linked list
implementation in `iteratorNext`.

Instead, OrderedHashMapBase keeps a list of active iterator indices
pointing to where traversal should visit next. This list is stored in a
ManagedChunkedList (`iteratorIndices_`). The design of
ManagedChunkedList allows the iterators to keep a pointer to the
iterator index stored in the list and can mark themselves as freed. The
ManagedChunkedList handles allocation and reuse of available slots.

The `newIterator` function will give the caller an `IteratorContext`.
This object stores the iterator index for the particular iterator that
holds the IteratorContext. Iterators will pass this context object to
the iterator functions, and those functions will make use of the
underlying index to perform traversal. The IteratorContext class is
also useful as a RAII class for use in JSMapImpl's `forEach` and
`forEachNative`.

The `advanceIterator` function is used to navigate to the next
undeleted element in the OrderedHashMapBase. And since the data table
is appended in insertion order, traversal is simply iterating to the
next set of values by updating the iterator index.

`iteratorKey` and `iteratorValue` are used to read from the data table
given the iterator index in the context object.

Reviewed By: lavenzg

Differential Revision: D81089296

fbshipit-source-id: 98c535d025fb2fec7d1180503137e7e4b4de17d0