Update data table when deleting entry
Summary:
When deleting an entry, we need to mark it accordingly in the data
table as well. This will become important for iterators later because
an iterator will traverse through the values in insertion order. It
needs some indicator for when to skip values.

So far the logic should all have zero impact on actual OrderedHashMap
functionality. They're done in parallel to the existing implementation.

Reviewed By: lavenzg

Differential Revision: D67180112

fbshipit-source-id: c913c118f0cd846e6b9676fab603ea326cdb7fb1