Use a RAII class to track StackOverflowGuard depths
Summary:
`StackOverflowGuard` has two types of heuristics. If
`HERMES_CHECK_NATIVE_STACK` is defined, it will use real stack checking
using the current stack address. Otherwise, it will use a simple depth
count.

In the latter, users of the stack overflow guard are expected to check
for `HERMES_CHECK_NATIVE_STACK`, then increment/decrement the depth
count if necessary.

This change introduces a simple RAII class that will handle this and
makes it simpler to use.

Reviewed By: avp

Differential Revision: D82996952

fbshipit-source-id: d058403c1d86f1a3ca77a5dacc14115e6641b5ef