Make WeakRefSlot support Symbol target
Summary:
Change WeakRoot to WeakSmallHermesValue in WeakRefSlot, so that it can
support Symbol target.

Note that now WeakRef<T>::get() runs one more step to get the result T*
on both HV64 and HV32 mode: it needs to get the CompressPointer first.
But this is just a bit masking operation so the cost is neglectable.

Reviewed By: avp

Differential Revision: D78119180

fbshipit-source-id: 934f23e27992ce0ba26b541bc91b3c21198e07d3