Make UnionNarrowTrustedInst unhoistable
Summary:
Mark UnionNarrowTrustedInst as unhoistable. This is to prevent
CodeMotion from moving the UnionNarrowTrustedInst above type checks. For
example, in TDZ there will first be a `ThrowIfInst` emitted, followed by
a `UnionNarrowTrustedInst`. The `UnionNarrowTrustedInst` is only vaild
in the non-throwing case of `ThrowIf`, but if the
`UnionNarrowTrustedInst` is lifted by CodeMotion, it would execute
before the verifying instruction of `ThrowIf`.

Reviewed By: avp

Differential Revision: D83385770

fbshipit-source-id: 3ce290b357c1d0b2989fc388d070f63a0bf0c707