Make GetConstructedObjectInst unhoistable
Summary:
Fix a bug with GetConstructedObjectInst by making
it unhoistable. The issue is that in some cases,
GetConstructedObjectInst may return a non-object type,
which violates its invariant. However, the only case that
can happen is when the Construct call which produced the
GetConstructedObjectInst will fail. In this case, it's
okay for the GetConstructedObjectInst to not return an object
because it was never going to run in the first place. However,
with CodeMotion, it can move the GetConstructedObjectInst
before the call, which would then break the invariant. So,
by setting the Unhoistable side effect, we prevent this issue.

Reviewed By: avp

Differential Revision: D83385768

fbshipit-source-id: d9556af121d08a85a39327faa6e5416790b7bfcf