Fix the test runner to search for directories
Summary:
The current test runner searches for just the test name when creating
the TestSuite for a test file. This could potentially be problematic if
the test file name contains the test suite name.

As a concrete example, the testrunner may look for the test
`xplat/third-party/javascript-test-suites/mjsunit/mjsunit.js` at a
non-existent path
`xplat/third-party/javascript-test-suites/mjsunit/mjsunit/mjsunit.js`

Fix the search to search for directories instead to prevent this.

Note that this was not caught because the problematic test files are in
the permanent blocklist currently. However, this was caught when trying
to update the mjsunit tests, which added more tests.

Reviewed By: lavenzg

Differential Revision: D85074330

fbshipit-source-id: 31dc93bd616cd20f5390ce1422a203f8ca0f03c7