Add failing shermes tests back to the skiplist
Summary:
We removed some mjsunit tests that are now passing from the skiplist.
These were tested with `hermes` on trunk. However, some tests may fail
in shermes only, so they cannot be removed yet. Adding the tests back.

Note: the mjsunits test did fail for shermes in the original diff, but
for some reason the results were skipped, so it was not caught.

Reviewed By: lavenzg

Differential Revision: D85181110

fbshipit-source-id: e2fb7a0b1e09ffa206727ccf863dd2ff88cd4cde