Create tag on release optimisically
Summary: Changes the `create-tag` to run immediately after the configuration of the run was set up without waiting for other steps to pass. It is for improving React Native releases to not wait for the tag creation. In case if anything goes wrong with the relase, the tag should be removed manually.

Reviewed By: cipolleschi

Differential Revision: D86193593

fbshipit-source-id: 1dcd62897e01f9814b58ff05edf16e622478f0bf