Fix jsi Instrumentation::getHeapInfo()
Summary:
`getHeapInfo()` in JSI is broken since we never populate
`youngGenStats` and `fullStats` in HadesGC. This diff fixed that.
While we are here, also remove `numMarkStackOverflows` in HeapInfo,
since we now do drain `barrierWorklist_` when it's full, for HadesGC
(this is done on mutator thread). For MallocGC, the mark stack is
always capped by the number of heap objects.

Also add a test to ensure we now do get some numbers in JSI
`getHeapInfo()`.

Reviewed By: avp

Differential Revision: D84677377

fbshipit-source-id: 401023037b76678214f8b635dc31a4c489203e71