Fix incorrect numCollections in GC stats
Summary:
`waitForCollectionToFinish()` is simply waiting for the ongoing OG
collection to finish, it should not increase the GC collection counter.
In printed GC stats, the total number of collections should be equal
to the number of YG collection + OG collection, otherwise it's
confusing.

This diff adds an additional parameter to `recordGCStats()` to indicate
whether it's called from a new collection cycle.

Reviewed By: avp

Differential Revision: D85736104

fbshipit-source-id: f5207252c497c122dc4ca524a4e62342b1c5e067