Handle columnless requests for breakpoints in lazy functions
Summary:
The `findSMLocFromCoords` function doesn't handle invalid columns
properly - it assumes that the column is real and searches for the line
and column.

In the debugger we use this to inform which lazy `CodeBlock`s we
actually compile in order to set the breakpoint.

When the user doesn't specify a column, we should instead compile any
`CodeBlock`s which are on the requested line. To facilitate this, add a
`findSMRangeForLine` which gives the start/end coordinates of the line,
and check the full range when deciding whether to compile the CodeBlock.

Reviewed By: tmikov

Differential Revision: D83595036

fbshipit-source-id: 0a48048442303727b45a3a9919fe3c38656bdb6d