Add integer fast path to scanNumber
Summary:
We already must scan over each character in the string representation of
the number, in order to find its bounds and construct input for
`fastStrToDouble`. It's trivial to also compute the value of the integer
during this traversal. Then, we can skip out on the call to
`fastStrToDouble` altogether.

Reviewed By: tmikov

Differential Revision: D86815226

fbshipit-source-id: 0f7136ca31b8529b7b85ecba93fdae883489403f