Refactor scanString
Summary:
Re-order code and make the fast path more important.

Break out of the fast path loop whenever an escape is
encountered. This is important so that we can maximize
the perf of escaped strings and non escaped strings
separately. When keeping them in one loop, improving the
perf of non escaped strings ended up having a dramatic
negative effect on escaped strings.

Reviewed By: avp

Differential Revision: D83611847

fbshipit-source-id: 6d1dd542f1c939426c61b5b3e1a83fadf80a18db