Use raw pointer type in JSONLexer
Summary:
Modify `JSONLexer` so that it operates directly on raw pointers when
lexing ASCII / UTF16 data. UTF8 data still must use UTF16Stream in order
to be lexed correctly. In order to do this as efficiently as posible,
the `JSONLexer` class is now templated on the input encoding type
(`EncodingKind`), which can be one of ASCII, UTF16, or UTF8. This
ensures that the code generated for advancing the current pointer into
the raw string is as best as it can be.

Reviewed By: avp

Differential Revision: D83270850

fbshipit-source-id: 4fbebb9b61630e2e61242555adb7ca0672477d71